requirements.txt
streamlit==1.41.1
faiss-cpu==1.9.0.post1
numpy==2.2.1
requests==2.32.3
sentence-transformers==5.1.1
pypdf==5.1.0
python-docx==1.1.2
openai==1.3.0
